+++ src/en/snowmtl/src/eu/kanade/tachiyomi/extension/en/snowmtl/Snowmtl.kt
@@ -1,13 +1,7 @@
 package eu.kanade.tachiyomi.extension.en.snowmtl
 
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import android.graphics.Canvas
-import android.graphics.Color
-import android.graphics.Paint
-import android.graphics.Rect
-import android.graphics.Typeface
 import eu.kanade.tachiyomi.network.GET
+import eu.kanade.tachiyomi.network.interceptor.rateLimit
 import eu.kanade.tachiyomi.source.model.Filter
 import eu.kanade.tachiyomi.source.model.FilterList
 import eu.kanade.tachiyomi.source.model.MangasPage
@@ -15,26 +9,18 @@ import eu.kanade.tachiyomi.source.model.Page
 import eu.kanade.tachiyomi.source.model.SChapter
 import eu.kanade.tachiyomi.source.model.SManga
 import eu.kanade.tachiyomi.source.online.ParsedHttpSource
+import kotlinx.serialization.decodeFromString
+import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonElement
-import kotlinx.serialization.json.JsonObject
-import kotlinx.serialization.json.jsonArray
-import kotlinx.serialization.json.jsonObject
-import kotlinx.serialization.json.jsonPrimitive
 import okhttp3.HttpUrl.Companion.toHttpUrl
-import okhttp3.Interceptor
-import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.Request
-import okhttp3.Response
-import okhttp3.ResponseBody.Companion.toResponseBody
 import org.jsoup.nodes.Document
 import org.jsoup.nodes.Element
 import rx.Observable
-import java.io.ByteArrayOutputStream
+import uy.kohesive.injekt.injectLazy
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.Locale
-import kotlin.math.absoluteValue
 
 class Snowmtl : ParsedHttpSource() {
 
@@ -46,11 +32,12 @@ class Snowmtl : ParsedHttpSource() {
 
     override val supportsLatest = true
 
-    override val client = network.cloudflareClient.newBuilder()
-        .addInterceptor(::composedImageIntercept)
-        .build()
+    private val json: Json by injectLazy()
 
-    private val captions = mutableMapOf<String, PageMapping>()
+    override val client = network.cloudflareClient.newBuilder()
+        .rateLimit(2)
+        .addInterceptor(ComposedImageInterceptor(baseUrl))
+        .build()
 
     // ============================== Popular ===============================
 
@@ -159,52 +146,25 @@ class Snowmtl : ParsedHttpSource() {
     // =============================== Pages ================================
 
     override fun pageListParse(document: Document): List<Page> {
-        captions.clear()
+        val pages = document.selectFirst("div#json-data")
+            ?.ownText()?.parseAs<List<PageDto>>()
+            ?: throw Exception("Pages not found")
 
-        val jsonArray = document.selectFirst("div#json-data")?.let {
-            Json.parseToJsonElement(it.ownText()).jsonArray
-        } ?: throw Exception("Pages not found")
-
-        return jsonArray.mapIndexed { index, jsonElement ->
-            val root = jsonElement.jsonObject
-            val imageUrl = """https://${root["img_url"]?.jsonPrimitive?.content}"""
-
-            captionsCache(root, imageUrl)
-
-            Page(index, imageUrl = imageUrl)
-        }
-    }
-
-    private fun captionsCache(root: JsonObject, imageUrl: String) {
-        val subs = root["translations"]!!.jsonArray.mapIndexed { index, it ->
-            val caption = try {
-                val arr = it.jsonArray
-                arr[0].jsonArray to arr[arr.size - 1].toContent()
-            } catch (_: Exception) {
-                val obj = it.jsonObject
-                obj["bbox"]!!.jsonArray to obj["text"]!!.toContent()
+        return pages.mapIndexed { index, dto ->
+            val imageUrl = when {
+                dto.imageUrl.startsWith("http") -> dto.imageUrl
+                else -> "https://${dto.imageUrl}"
             }
-
-            val coord = caption.first
-
-            Sub(
-                index = index,
-                x1 = coord[0].toInt(),
-                y1 = coord[1].toInt(),
-                x2 = coord[2].toInt(),
-                y2 = coord[3].toInt(),
-                text = caption.second,
-            )
+            val fragment = json.encodeToString<List<Translation>>(dto.translations)
+            Page(index, imageUrl = "$imageUrl#$fragment")
         }
-
-        captions[imageUrl] = PageMapping(url = imageUrl, subs)
     }
 
     override fun imageUrlParse(document: Document): String = ""
 
     // ============================= Utilities ==============================
 
-    fun parseChapterDate(date: String?): Long {
+    private fun parseChapterDate(date: String?): Long {
         date ?: return 0
         return try { dateFormat.parse(date)!!.time } catch (_: Exception) { parseRelativeDate(date) }
     }
@@ -214,179 +174,17 @@ class Snowmtl : ParsedHttpSource() {
         val cal = Calendar.getInstance()
 
         return when {
-            date.contains("day") -> cal.apply { add(Calendar.DAY_OF_MONTH, -number) }.timeInMillis
-            date.contains("hour") -> cal.apply { add(Calendar.HOUR, -number) }.timeInMillis
-            date.contains("minute") -> cal.apply { add(Calendar.MINUTE, -number) }.timeInMillis
-            date.contains("second") -> cal.apply { add(Calendar.SECOND, -number) }.timeInMillis
-            date.contains("week") -> cal.apply { add(Calendar.DAY_OF_MONTH, -number * 7) }.timeInMillis
+            date.contains("day", true) -> cal.apply { add(Calendar.DAY_OF_MONTH, -number) }.timeInMillis
+            date.contains("hour", true) -> cal.apply { add(Calendar.HOUR, -number) }.timeInMillis
+            date.contains("minute", true) -> cal.apply { add(Calendar.MINUTE, -number) }.timeInMillis
+            date.contains("second", true) -> cal.apply { add(Calendar.SECOND, -number) }.timeInMillis
+            date.contains("week", true) -> cal.apply { add(Calendar.DAY_OF_MONTH, -number * 7) }.timeInMillis
             else -> 0
         }
     }
 
-    private fun JsonElement.toInt() = this.jsonPrimitive.content.toInt()
-
-    private fun JsonElement.toContent() = this.jsonPrimitive.content
-
-    val Int.sp: Float get() = this * scaledDensity
-
-    // The Interceptor joins the subtitles and chapters of the manga.
-    private fun composedImageIntercept(chain: Interceptor.Chain): Response {
-        val request = chain.request()
-        val url = request.url.toString()
-        if ((url in captions).not()) {
-            return chain.proceed(request)
-        }
-
-        val response = chain.proceed(request)
-
-        val bitmap = BitmapFactory.decodeStream(response.body.byteStream())!!
-            .copy(Bitmap.Config.ARGB_8888, true)
-
-        val canvas = Canvas(bitmap)
-        val defaultTextSize = 22.sp // random
-
-        val paint = Paint().apply {
-            textAlign = Paint.Align.LEFT
-            color = Color.BLACK
-            style = Paint.Style.FILL_AND_STROKE
-            textSize = defaultTextSize
-            isAntiAlias = true
-            typeface = Typeface.SANS_SERIF
-        }
-
-        val marginTop = 30 // random
-        val marginLeft = 30 // random
-        val spaceBetween = 2 // random
-        val defaultFontScale = 1f
-
-        captions[url]?.subs
-            ?.filter { it.text.isNotBlank() }
-            ?.forEach {
-                var lines = it.breakLines(paint)
-                /*
-                    Reduces the font according to the size of the line in the dialog box. (space between applied)
-                        Ex. 1:
-                            - Box: 10 lines
-                            - Text: 9 lines
-                            - Scale: 1
-                        Ex. 2:
-                            - Box: 10 lines
-                            - Text: 15
-                            - Scale: 0.6 (10/ (15 + 1)) // 1 extra line(random)
-
-                        Ex. 3:
-                            - Box: 10 lines
-                            - Text: 2 line
-                            - Scale: 1
-                 */
-                val dialogBoxLines = it.height / paint.getCharHeight()
-                val fontScale = when {
-                    lines.size >= dialogBoxLines -> dialogBoxLines / (lines.size + 1)
-                    else -> defaultFontScale
-                }
-
-                // Use font scale in large dialogs
-                if (fontScale != defaultFontScale) {
-                    paint.apply {
-                        this.textSize = defaultTextSize * fontScale
-                    }
-                    // reprocessing break lines
-                    lines = it.breakLines(paint)
-                }
-
-                // Centers the text if it is smaller than half of the dialog box.
-                val isHalfTheBox = lines.size / dialogBoxLines < 0.5
-                val initialY = when {
-                    isHalfTheBox -> it.centerY - lines.size * paint.getCharHeight() / 2
-                    else -> it.y1.toFloat()
-                }
-
-                // Invert color in black dialog box
-                paint.apply {
-                    val pixelColor = bitmap.getPixel(it.centerX.toInt(), it.centerY.toInt())
-                    val inverseColor = (Color.WHITE - pixelColor) or Color.BLACK
-                    color = inverseColor
-                }
-
-                lines.forEachIndexed { index, line ->
-                    // Centers the text on the X axis and positions it inside the dialog box
-                    val x = (it.centerX - (line.length * paint.getCharWidth() / 2)).absoluteValue + marginLeft
-
-                    // Positions the text inside the dialog box on the Y axis
-                    val y = (initialY + paint.getCharHeight() * index * spaceBetween).absoluteValue + marginTop
-
-                    canvas.drawText(line, 0, line.length, x, y, paint)
-                }
-            }
-
-        val output = ByteArrayOutputStream()
-
-        val format = when (url.substringAfterLast(".").lowercase()) {
-            "png" -> Bitmap.CompressFormat.PNG
-            "jpeg", "jpg" -> Bitmap.CompressFormat.JPEG
-            else -> Bitmap.CompressFormat.WEBP
-        }
-
-        bitmap.compress(format, 100, output)
-
-        val responseBody = output.toByteArray().toResponseBody(MEDIA_TYPE)
-
-        return response.newBuilder()
-            .body(responseBody)
-            .build()
-    }
-
-    data class PageMapping(
-        val url: String,
-        val subs: List<Sub>,
-    )
-
-    data class Sub(
-        val index: Int,
-        val x1: Int,
-        val y1: Int,
-        val x2: Int,
-        val y2: Int,
-        val text: String,
-    ) {
-        val width get() = x2 - x1
-        val height get() = y2 - y1
-        val centerY get() = (y2 + y1) / 2f
-        val centerX get() = (x2 + x1) / 2f
-
-        fun breakLines(paint: Paint): List<String> {
-            val diameter = width / paint.getCharWidth()
-            val radius = diameter / 2
-            return breakTextIntoLines(text, diameter + radius)
-        }
-
-        private fun breakTextIntoLines(text: String, maxLineLength: Float): List<String> {
-            if (text.length <= maxLineLength) {
-                return listOf(text)
-            }
-
-            val words = text.split(" ")
-            val lines = mutableListOf<String>()
-            var currentLine = StringBuilder()
-
-            for (word in words) {
-                if (currentLine.length + word.length + 1 <= maxLineLength) {
-                    if (currentLine.isNotEmpty()) {
-                        currentLine.append(" ")
-                    }
-                    currentLine.append(word)
-                } else {
-                    lines.add(currentLine.toString().trim())
-                    currentLine = StringBuilder(word)
-                }
-            }
-
-            if (currentLine.isNotEmpty()) {
-                lines.add(currentLine.toString())
-            }
-
-            return lines
-        }
+    private inline fun <reified T> String.parseAs(): T {
+        return json.decodeFromString(this)
     }
 
     // =============================== Filters ================================
@@ -401,80 +199,8 @@ class Snowmtl : ParsedHttpSource() {
         return FilterList(filters)
     }
 
-    private var genreList: List<Genre> = listOf(
-        Genre("Action"),
-        Genre("Adult"),
-        Genre("Adventure"),
-        Genre("Comedy"),
-        Genre("Drama"),
-        Genre("Ecchi"),
-        Genre("Fantasy"),
-        Genre("Gender Bender"),
-        Genre("Harem"),
-        Genre("Historical"),
-        Genre("Horror"),
-        Genre("Josei"),
-        Genre("Lolicon"),
-        Genre("Martial Arts"),
-        Genre("Mature"),
-        Genre("Mecha"),
-        Genre("Mystery"),
-        Genre("Psychological"),
-        Genre("Romance"),
-        Genre("School Life"),
-        Genre("Sci-fi"),
-        Genre("Seinen"),
-        Genre("Shoujo"),
-        Genre("Shoujo Ai"),
-        Genre("Shounen"),
-        Genre("Shounen Ai"),
-        Genre("Slice of Life"),
-        Genre("Smut"),
-        Genre("Sports"),
-        Genre("Supernatural"),
-        Genre("Tragedy"),
-        Genre("Yaoi"),
-        Genre("Yuri"),
-    )
-
-    private val sortByList = listOf(
-        Option("All"),
-        Option("Most Views", "views"),
-        Option("Most Recent", "recent"),
-    ).map { it.copy(query = "sort_by") }
-
-    private data class Option(val name: String = "", val value: String = "", val query: String = "")
-
-    private open class SelectionList(displayName: String, private val vals: List<Option>, state: Int = 0) :
-        Filter.Select<String>(displayName, vals.map { it.name }.toTypedArray(), state) {
-        fun selected() = vals[state]
-    }
-
-    private class GenreList(title: String, genres: List<Genre>) :
-        Filter.Group<GenreCheckBox>(title, genres.map { GenreCheckBox(it.name, it.id) })
-
-    class GenreCheckBox(name: String, val id: String = name) : Filter.CheckBox(name)
-
-    class Genre(val name: String, val id: String = name)
-
     companion object {
-        const val scaledDensity = 1.5f
         const val PREFIX_SEARCH = "id:"
-        val MEDIA_TYPE = "image/png".toMediaType()
         private val dateFormat: SimpleDateFormat = SimpleDateFormat("dd MMMM yyyy", Locale.US)
     }
-}
-
-fun Paint.getCharWidth(): Float {
-    val text = "A"
-    val fontWidth = FloatArray(1)
-    getTextWidths(text.first().toString(), fontWidth)
-    return fontWidth.first()
-}
-
-fun Paint.getCharHeight(): Float {
-    val text = "A"
-    val bounds = Rect()
-    getTextBounds(text, 0, text.length, bounds)
-    return bounds.height().toFloat()
 }
